requirements.txt
captum==0.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
clip==1.0
contourpy==1.0.5
cycler==0.11.0
einops==0.5.0
filelock==3.8.0
fonttools==4.37.4
ftfy==6.1.1
huggingface-hub==0.10.1
idna==3.4
imageio==2.22.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pycountry==22.3.5
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.1
-e git+https://github.com/huggingface/transformers.git@072ed01c388ee04a452619c57ed40bb97f5e9953#egg=transformers
typing-extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
